refactor: isolate diagnostics environment capture
